% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setEnvironment.R
\docType{data}
\name{cacheEnv}
\alias{cacheEnv}
\title{cacheEnv Package variable available to all functions in package}
\format{An object of class \code{environment} of length 0.}
\usage{
cacheEnv
}
\description{
cacheEnv Package variable available to all functions in package
}
\keyword{datasets}
