% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSWeb.R
\docType{package}
\name{CSWeb}
\alias{CSWeb}
\alias{CSWeb-package}
\title{CSWeb: A package for Credit Suisse for Web Access/Scraping .}
\description{
The foo package provides functions to access Web Content
by passing the required URL to getData function.
#'
}
